+++ database/migrations/2023_03_10_090811_create_orders_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->foreignId('user_id');
             $table->string('payment_intent_id')->nullable();        // nullable if the payment is made with cash
             $table->string('payment_type');
-            $table->integer('total_price');
+            $table->float('total_price');
             $table->text('description')->nullable();
             $table->timestamps();
